app/requirements.txt
flask
psycopg2-binary
redis
prometheus-client

